feat: Add VecBdlcs which uses interior mutability
VecBdlcs is a vector of BiDirLocalChannels that uses UnsafeCell directly to create
interior mutability of the vector. I tried using RefCell but when dereferencing a RefCell
it returns Ref and I ended up having a temporary that didn't live long enough. This
is "working" but obviously using Unsafe isn't the best!

 - Enabled `selector.recv(x.our_channel.get_recv())`
 - Passing rsp_tx to proceess_msg_any, previously was passing `None`
 - Rename "Get" to "Req" and "Reply" to "Rsp" (both are 3 characters)
 - Other clean up